Move static plugin loading methods to util class
We had a set of utility methods in PluginsService that were public and
static so that the plugin cli tool could use them. Here, we move those
methods to a utility class so that the plugin CLI doesn't need to import
PluginsService. For the same reason, we also move
"PluginsService.Bundle" to a top-level class called "PluginBundle."
Finally, we move tests to the appropriate test classes.
