+++ distribution/tools/plugin-cli/src/main/java/org/elasticsearch/plugins/cli/InstallPluginAction.java
@@ -39,8 +39,9 @@ import org.elasticsearch.core.Tuple;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.jdk.JarHell;
 import org.elasticsearch.plugins.Platforms;
+import org.elasticsearch.plugins.PluginBundle;
 import org.elasticsearch.plugins.PluginInfo;
-import org.elasticsearch.plugins.PluginsService;
+import org.elasticsearch.plugins.PluginsUtils;
 
 import java.io.BufferedReader;
 import java.io.Closeable;
@@ -851,12 +852,12 @@ public class InstallPluginAction implements Closeable {
         if (info.hasNativeController()) {
             throw new IllegalStateException("plugins can not have native controllers");
         }
-        PluginsService.verifyCompatibility(info);
+        PluginsUtils.verifyCompatibility(info);
 
         // checking for existing version of the plugin
         verifyPluginName(env.pluginsFile(), info.getName());
 
-        PluginsService.checkForFailedPluginRemovals(env.pluginsFile());
+        PluginsUtils.checkForFailedPluginRemovals(env.pluginsFile());
 
         terminal.println(VERBOSE, info.toString());
 
@@ -886,16 +887,16 @@ public class InstallPluginAction implements Closeable {
         }).collect(Collectors.toSet());
 
         // read existing bundles. this does some checks on the installation too.
-        Set<PluginsService.Bundle> bundles = new HashSet<>(PluginsService.getPluginBundles(pluginsDir));
-        bundles.addAll(PluginsService.getModuleBundles(modulesDir));
-        bundles.add(new PluginsService.Bundle(candidateInfo, candidateDir));
-        List<PluginsService.Bundle> sortedBundles = PluginsService.sortBundles(bundles);
+        Set<PluginBundle> bundles = new HashSet<>(PluginsUtils.getPluginBundles(pluginsDir));
+        bundles.addAll(PluginsUtils.getModuleBundles(modulesDir));
+        bundles.add(new PluginBundle(candidateInfo, candidateDir));
+        List<PluginBundle> sortedBundles = PluginsUtils.sortBundles(bundles);
 
         // check jarhell of all plugins so we know this plugin and anything depending on it are ok together
         // TODO: optimize to skip any bundles not connected to the candidate plugin?
         Map<String, Set<URL>> transitiveUrls = new HashMap<>();
-        for (PluginsService.Bundle bundle : sortedBundles) {
-            PluginsService.checkBundleJarHell(classpath, bundle, transitiveUrls);
+        for (PluginBundle bundle : sortedBundles) {
+            PluginsUtils.checkBundleJarHell(classpath, bundle, transitiveUrls);
         }
 
         // TODO: no jars should be an error
